merge(hangul): merge internal changes
