+++ packages/common/hangul/src/disassemble.en.md
@@ -1,3 +1,7 @@
+---
+title: disassemble
+---
+
 # disassembleHangulToGroups
 
 Disassemble a Hangul string into syllables, but group them by its original character.
@@ -20,4 +24,26 @@ disassembleHangulToGroups('값'); // [['ㄱ', 'ㅏ', 'ㅂ', 'ㅅ']]
 disassembleHangulToGroups('토스 짱'); // [['ㅌ', 'ㅗ'], ['ㅅ', 'ㅡ'], [' '], ['ㅉ', 'ㅏ', 'ㅇ']]
 disassembleHangulToGroups('ㅘ'); // [['ㅗ', 'ㅏ']]
 disassembleHangulToGroups('ㄵ'); // [['ㄴ', 'ㅈ']]
+```
+
+# disassembleHangul
+
+Disassemble a hangul string into first/middle/last letters, making a single string.
+
+See below for detailed examples.
+
+```typescript
+function disassembleHangul(
+  // The hangul string to disassemble
+  str: string
+): string;
+```
+
+## Examples
+
+```tsx
+disassembleHangul('값'); // 'ㄱㅏㅂㅅ'
+disassembleHangul('값이 비싸다'); // 'ㄱㅏㅂㅅㅇㅣ ㅂㅣㅆㅏㄷㅏ'
+disassembleHangul('ㅘ'); // 'ㅗㅏ'
+disassembleHangul('ㄵ'); // 'ㄴㅈ'
 ```
