+++ packages/common/hangul/src/disassemble.ko.md
@@ -1,3 +1,7 @@
+---
+title: disassemble
+---
+
 # disassembleHangulToGroups
 
 한글 문자열을 글자별로 초성/중성/종성 단위로 완전히 분리합니다.
@@ -20,4 +24,26 @@ disassembleHangulToGroups('값'); // [['ㄱ', 'ㅏ', 'ㅂ', 'ㅅ']]
 disassembleHangulToGroups('토스 짱'); // [['ㅌ', 'ㅗ'], ['ㅅ', 'ㅡ'], [' '], ['ㅉ', 'ㅏ', 'ㅇ']]
 disassembleHangulToGroups('ㅘ'); // [['ㅗ', 'ㅏ']]
 disassembleHangulToGroups('ㄵ'); // [['ㄴ', 'ㅈ']]
+```
+
+# disassembleHangul
+
+한글 문자열을 글자별로 초성/중성/종성 단위로 완전히 분리하여, 하나의 문자열로 만듭니다.
+
+자세한 예시는 아래 Example을 참고하세요.
+
+```typescript
+function disassembleHangul(
+  // 분리할 한글 문자열
+  str: string
+): string;
+```
+
+## Examples
+
+```tsx
+disassembleHangul('값'); // 'ㄱㅏㅂㅅ'
+disassembleHangul('값이 비싸다'); // 'ㄱㅏㅂㅅㅇㅣ ㅂㅣㅆㅏㄷㅏ'
+disassembleHangul('ㅘ'); // 'ㅗㅏ'
+disassembleHangul('ㄵ'); // 'ㄴㅈ'
 ```
