+++ packages/common/hangul/src/disassemble.spec.ts
@@ -1,4 +1,4 @@
-import { disassembleHangulToGroups } from './disassembleHangulToGroups';
+import { disassembleHangulToGroups, disassembleHangul } from './disassemble';
 
 describe('disassembleHangulToGroups', () => {
   it('값', () => {
@@ -26,5 +26,27 @@ describe('disassembleHangulToGroups', () => {
 
   it('ㅘ', () => {
     expect(disassembleHangulToGroups('ㅘ')).toEqual([['ㅗ', 'ㅏ']]);
+  });
+});
+
+describe('disassembleHangul', () => {
+  it('값', () => {
+    expect(disassembleHangul('값')).toEqual('ㄱㅏㅂㅅ');
+  });
+
+  it('값이 비싸다', () => {
+    expect(disassembleHangul('값이 비싸다')).toEqual('ㄱㅏㅂㅅㅇㅣ ㅂㅣㅆㅏㄷㅏ');
+  });
+
+  it('토스 짱', () => {
+    expect(disassembleHangul('토스 짱')).toEqual('ㅌㅗㅅㅡ ㅉㅏㅇ');
+  });
+
+  it('ㄵ', () => {
+    expect(disassembleHangul('ㄵ')).toEqual('ㄴㅈ');
+  });
+
+  it('ㅘ', () => {
+    expect(disassembleHangul('ㅘ')).toEqual('ㅗㅏ');
   });
 });
