+++ packages/common/hangul/src/disassemble.ts
@@ -1,4 +1,3 @@
-/** @tossdocs-ignore */
 import { DISASSEMBLED_CONSONANTS_BY_CONSONANT, DISASSEMBLED_VOWELS_BY_VOWEL } from './constants';
 import { disassembleCompleteHangulCharacter } from './disassembleCompleteHangulCharacter';
 
@@ -38,4 +37,8 @@ export function disassembleHangulToGroups(str: string) {
   }
 
   return result;
+}
+
+export function disassembleHangul(str: string) {
+  return disassembleHangulToGroups(str).reduce((hanguls, disassembleds) => `${hanguls}${disassembleds.join('')}`, '');
 }
