+++ packages/common/hangul/src/hangulIncludes.ts
@@ -1,5 +1,4 @@
-/** @tossdocs-ignore */
-import { disassembleHangul } from './disassembleHangul';
+import { disassembleHangul } from './disassemble';
 
 export function hangulIncludes(x: string, y: string) {
   const disassembledX = disassembleHangul(x);
